Added Protection

+++ app.js
@@ -49,7 +49,7 @@ app.post('/login',async (req,res)=>{
     bcrypt.compare(req.body.password, user.password, function(err, result) {
         if (result) {
             
-            let token = jwt. sign({email:`${user.email}`},'Secret')
+            let token = jwt.sign({email: user.email, id: user._id}, "Secret")
             res.cookie("token",token)
             res.redirect('/userSitePage')
         } else {
@@ -89,6 +89,7 @@ function checkAuth(req, res, next) {
 //Added
 
 app.get('/editProfile/:id',checkAuth, async (req,res)=>{
+    if (req.user.id !== req.params.id) return res.status(403).send("Forbidden")
     let user = await userModel.findOne({_id:req.params.id}) //We are using findOne here so we can return only an object . Thus, no destructuring is needed \
     // {user} unlike the previous /read route where {users:allusers} had to be implemented
 
@@ -97,12 +98,14 @@ app.get('/editProfile/:id',checkAuth, async (req,res)=>{
 
 app.post('/update/:userid', checkAuth, async(req,res)=>
 {
+    if (req.user.id !== req.params.id) return res.status(403).send("Forbidden")
     let {username,email,image} = req.body //Stores the updated (if any) record
     let user = await userModel.findOneAndUpdate({_id:req.params.userid}, {username,image}, {new:true}) // Updates the user
     res.redirect('/userSitePage') // Redirects to the read page
 })
 
 app.get('/deleteProfile/:ID',checkAuth, async(req,res)=>{
+    if (req.user.id !== req.params.id) return res.status(403).send("Forbidden")
     const {ID} = req.params // req.params is used only when dynamic routing is applied. Otherwise, any thing dealing with the frontend aspect will normally
     //result in req.body as seen in the /create route where the client enters
 
@@ -114,6 +117,7 @@ app.get('/deleteProfile/:ID',checkAuth, async(req,res)=>{
 
 app.get('/readFiles/:id',checkAuth,(req,res)=>
     {
+        if (req.user.id !== req.params.id) return res.status(403).send("Forbidden")        
         const id = req.params.id
         fs.readdir(`./Folders/${id}`,function(err,files){
             res.render('index_updt',{files: files, id:id})
@@ -123,6 +127,7 @@ app.get('/readFiles/:id',checkAuth,(req,res)=>
 
 app.post('/createFile/:id',checkAuth,(req,res)=>
     { 
+        if (req.user.id !== req.params.id) return res.status(403).send("Forbidden")
         const id = req.params.id
         //Writing a file into the directory which will be retrieved in the frontend aspect + req.body reason is because the title,description are under that HTML aspect
         fs.writeFile(`./Folders/${id}/${req.body.title.split(' ').join('')}.txt`,req.body.description,function(err){
@@ -131,6 +136,7 @@ app.post('/createFile/:id',checkAuth,(req,res)=>
     })
 
 app.get('/fileView/:id/:fileName', checkAuth,(req, res) => {
+  if (req.user.id !== req.params.id) return res.status(403).send("Forbidden")
   const { id, fileName } = req.params
   fs.readFile(`./Folders/${id}/${fileName}`, 'utf8', (err, data) => {
     if (err) {console.error(err)
@@ -140,6 +146,7 @@ app.get('/fileView/:id/:fileName', checkAuth,(req, res) => {
 });
 
 app.get('/fileEdit/:id/:fileName',checkAuth, (req,res)=>{
+    if (req.user.id !== req.params.id) return res.status(403).send("Forbidden")
     const {id,fileName} = req.params
     fs.readFile(`./Folders/${id}/${fileName}`,'utf-8',(err,data)=>{
         if(err){console.error(err) 
@@ -150,6 +157,7 @@ app.get('/fileEdit/:id/:fileName',checkAuth, (req,res)=>{
 })
 
 app.post('/updateFile/:id',checkAuth, (req,res)=>{
+    if (req.user.id !== req.params.id) return res.status(403).send("Forbidden") // Simple Protection
     const {id} = req.params
     const fileName = req.body.title.split(' ').join('') + '.txt'
     fs.writeFile(`./Folders/${id}/${req.body.title.split(' ').join('')}.txt`,req.body.description,function(err){ res.redirect(`/fileView/${id}/${fileName}`)}
